Move bindings for FilterSource into the common module
