+++ examples/digital_fingerprinting/production/morpheus/dfp_duo_pipeline.py
@@ -46,8 +46,8 @@ from dfp.utils.column_info import create_increment_col
 from dfp.utils.file_utils import date_extractor
 from dfp.utils.file_utils import iso_date_regex
 
+from morpheus._lib.common import FilterSource
 from morpheus._lib.file_types import FileTypes
-from morpheus._lib.filter_source import FilterSource
 from morpheus.cli.utils import get_package_relative_file
 from morpheus.cli.utils import load_labels_file
 from morpheus.config import Config
