+++ morpheus/_lib/CMakeLists.txt
@@ -54,9 +54,6 @@ include(cmake/python_modules/common.cmake)
 #----------morpheus._lib.file_types---------
 include(cmake/python_modules/file_types.cmake)
 
-#----------morpheus._lib.filter_source---------
-include(cmake/python_modules/filter_source.cmake)
-
 if (MORPHEUS_BUILD_TESTS)
   add_subdirectory(tests)
 endif()
