--- morpheus/_lib/cmake/python_modules/filter_source.cmake
@@ -1,34 +1,0 @@
-# =============================================================================
-# Copyright (c) 2020-2023, NVIDIA CORPORATION.
-#
-# Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
-# in compliance with the License. You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software distributed under the License
-# is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
-# or implied. See the License for the specific language governing permissions and limitations under
-# the License.
-# =============================================================================
-
-morpheus_add_pybind11_module(
-  filter_source
-    MODULE_ROOT
-      "${MORPHEUS_LIB_ROOT}"
-    SOURCE_FILES
-      "${MORPHEUS_LIB_ROOT}/src/python_modules/filter_source.cpp"
-    INCLUDE_DIRS
-      "${MORPHEUS_LIB_ROOT}/include"
-    LINK_TARGETS
-      morpheus
-      mrc::pymrc
-    OUTPUT_TARGET
-      filter_source_target
-    INSTALL_DEST
-      ${MORPHEUS_LIB_INSTALL_DIR}
-)
-
-if(MORPHEUS_PYTHON_INPLACE_BUILD)
-  inplace_build_copy(${filter_source_target} ${MORPHEUS_LIB_ROOT})
-endif()
