+++ morpheus/_lib/src/python_modules/stages.cpp
@@ -53,7 +53,7 @@ PYBIND11_MODULE(stages, m)
     // Load the cudf helpers
     load_cudf_helpers();
 
-    mrc::pymrc::from_import(m, "morpheus._lib.filter_source", "FilterSource");
+    mrc::pymrc::from_import(m, "morpheus._lib.common", "FilterSource");
 
     py::class_<mrc::segment::Object<AddClassificationsStage>,
                mrc::segment::ObjectProperties,
