+++ morpheus/stages/postprocess/filter_detections_stage.py
@@ -21,7 +21,7 @@ import numpy as np
 from mrc.core import operators as ops
 
 import morpheus._lib.stages as _stages
-from morpheus._lib.filter_source import FilterSource
+from morpheus._lib.common import FilterSource
 from morpheus.cli.register_stage import register_stage
 from morpheus.config import Config
 from morpheus.messages import MultiMessage
@@ -68,7 +68,7 @@ class FilterDetectionsStage(SinglePortStage):
         Threshold to classify, default is 0.5.
     copy : bool
         Whether or not to perform a copy.
-    filter_source : `morpheus._lib.filter_source.FilterSource`, default = 'auto'
+    filter_source : `from morpheus._lib.common.FilterSource`, default = 'auto'
         Indicate if we are operating on is an output tensor or a field in the DataFrame.
         Choosing `Auto` will default to `TENSOR` and in a future release will change to `DATAFRAME`
     field_name : str
